Preserve isNetworkIdentifiedEmergencyCall for Connection during handover.
- During a handover, the Connection object is overwritten.
This can cause isNetworkIdentifiedEmergencyCall to be false.
The Call object will never reset its boolean member for emergency
calls. Once the call is an emergency call, that never changes from
the point of view of the Call object. This can lead to a discrepency
between the Call and Connection. Caching isNetworkIdentifiedEmergencyCall
resolves this issue.

- Make sure network identified emergency calls are treated as emergency
calls so the UI is consistent for both normal and network identified
emergency call (e.g. the hold button is not shown).

Bug : 343692509
Flag : EXEMPT bugfix
Test : UT verified

Change-Id: Iff8bbc469879de151c1709c900a16325718928e4