DSDA: Wait for call disconnect on other subs for ECC
Implements waiting logic in Telephony to disconnect calls on the other
subs when we try to place an emergency call. Performs a bulk transaction
and completes all the disconnects irregardless of individual failures.

This will help ensure that we don't have calls active/held across sims
when simultaneous calling is available and also address the shared data
call issue with not waiting for the disconnect to occur on the shared
data sim before placing the emergency call on the sim in service.

Bug: 392910450
Bug: 393240400
Flag: com.android.server.telecom.flags.enable_call_sequencing
Test: atest TelephonyConnectionServiceTest
Change-Id: I536d7334d908f0a7f349e10228bd08e188c621ae