Ensure the VoicemailSettingsActivity supports edge to edge.
The PreferenceActivity itself needs to be configured to support edge to
edge properly.

Also, it turns out this preference activity is unique in that it nests
PreferenceScreens.  As a result the "voicemail number" which is nested
in the "Setup" PreferenceScreen would get cut off.

Flag: NONE bug fix.
Fixes: 394283757
Fixes: 398759149
Test: Manual test of the telephony voicemail screen to demonstrate that
edge to edge does not break the layout.

Change-Id: I4410d184700b9fff5cd54d3e716c127a4f762596