Use edge to edge listener on phone account settings activity.
It has been reported that sometimes the calling accounts screen will
revert to a non edge-to-edge compliant layout.  I am unable to repro this
so I will try using the "setupEdgeToEdge" which uses a listener
internally to handle this case.  Perhaps that'll be more reliable.
I suspect perhaps it is to do with the fact that this screen is hosted in
a fragment?  Don't know.  I left the `getView().setFitsSystemWindows(true)`
in the fragment itself because that sometimes gets loaded from
`CallFeatureSettings` in some scenarios.  Usingboth does not seem to cause
any harm.

Flag: NONE bugfix
Test: View screen to confirm edge to edge works as expected.
Bug: 392617692
Change-Id: I60eaa2048fec179c7a9d086138b8b84791cf8b3b