DSDA: Cleanup outgoing ECC logic
Ensure that we hold call on same sub before dialing out the ECC. It's
possible in some cases that we don't ensure that the active call on the
same sub that we're placing the ECC is not held before. Also move the
disconnects on the other subs to be synchronous as they were causing
unnecessary delays in the disconnect, so the incall screen for the
disconnected call would remain when the ECC call was answered.

Bug: 403620519
Flag: com.android.server.telecom.flags.enable_call_sequencing
Test: manual verification of ECC test cases with multiple calls,
emulating both the non-holdable case as well as carriers that don't
support emergency call holding
Test: atest CtsTelecomCujTestCases

Change-Id: I9c86521f8d9061cc11d3bd513f85e93dbed0f5e7