Dedupe voicemail broadcasts sent to the default Dialer.
The voicemail message waiting broadcasts can potentially be sent in
duplicate MANY MANY times on bootup.  This stems from the fact that
the MWI indication is updated both when the network sends up signalling
to say it has happened, AND because any changes to SubscriptionInfo can
trigger a "refresh".  In practice I saw up to 6 broadcasts at bootup
for the same thing -- this means that Dialer is likely to blow past the
10 seconds it has to process the flury of broadcasts and ANR.  This was
likely not an issue in the past because there were less attributes in
subinfo.  Something new in there must be triggering sub info changes more
often now, which causes a lot of extra chatter.

The actual change here is in NotificationMgr; the rest of extra comments
and logs to help understand what is going on here.  This is old code and
the way it works is not very intuitive, so the hope is that this will at
the very least will be a bit easier to debug in the future.

Test: Manual testing on live network by receiving and clearing voicemail
messages, rebooting, etc.
Flag: NONE bugfix
Fixes: 376195239

Change-Id: Ia43ff1dd70a93d57b8accb390e2bc50dcb93b586