Resolve UT error on ImsConferenceControllerTest.
This CL fixes a UT error in ImsCOnferenceControllerTest#testMergeMultiPartyCalls. That test was failing when the new addConferenceFromConnection behavior was enabled. This CL updates the test to test for the new expected behavior and adds a legacy test to test the behavior when the relavant flag is disabled.

Flag: com.android.internal.telephony.flags.reuse_original_conn_remote_conf_behavior
Bug: 391889544
Test: ImsConferenceControllerTest
Change-Id: I184bc18bfc1669cc30f696d2f7c9eea057726fba