[RemoteSIM]: Check for config before throwing exception.
This follows the precedent elsewhere inside PhoneInterfaceManager for
handling this config. The config_force_phone_globals_creation config
is enabled on devices that don't have access to all of Telephony but
use go/remote-sim. This is now required on Android desktop to support
messaging for go/al-beto-ios-hub-dd.

Test: existing tests pass
Flag: EXEMPT bugfix
Bug: b/421008371, b/421015000
Change-Id: I540c4fabd64f18514d51d7a613e8ba696ea7fd2d