Add logic to control holding RTT calls via carrier config in Telephony.
This change introduces the logic to control whether an RTT call can be
placed on hold based on the KEY_ALLOW_HOLD_IN_RTT_CALL_BOOL carrier
config.

Design Doc: go/rtt-hold-unhold-config-in-android-connectivity-design-doc

Additionally, this CL fixes a bug where the hold capability was not
correctly restored after an RTT session ended. This was because the
connection capabilities were not being updated after RTT termination.
The fix is to call updateConnectionCapabilities() when an RTT session is
initiated or terminated.

Bug: 413765645
Test: Live network test video: go/rtt-hold-live-network-test
Test: atest ImsCallingTest TeleServiceTests
Flag: com.android.internal.telephony.flags.enable_rtt_hold_carrier_config
Change-Id: Ia8556cea05a02b7fa5002a46041f10ded4d153a4