[RemoteSIM]: Check for config before throwing exception.
This follows the precedent elsewhere inside PhoneInterfaceManager for
handling this config. The config_force_phone_globals_creation config
is enabled on devices that don't have access to all of Telephony but
use go/remote-sim. This is now required on Android desktop to support
messaging for go/al-beto-ios-hub-dd.

Test: existing unittests pass
Bug: b/421015000
Flag: EXEMPT bugfix
Change-Id: I6ccf0b7dea20ff4b2bcc5135810128682e1af8d3