Delay connection state details update during domain reselection
This change is a fix to prevent incorrect updates of connection property
information managed by the Telecom layer during domain reselection,
which causes the state details of the connection to be updated.
The change is to maintain consistency with the behavior of the devices
with AP-DS disabled by updating the state details of the connection when
a call is disconnected or a call redial occurs after domain reselection.

The issue occurs because the RTT property is removed during domain
reselection, making it impossible to determine whether the currently
ended call was an RTT call when an RTT emergency call fails.

Bug: 429443154
Flag: com.android.internal.telephony.flags.ignore_state_details_update_for_domain_reselection
Test: atest TelephonyConnectionTest, TelephonyConnectionServiceTest
Test: manual (verified voice emergency call is redialed after RTT
emergency call failed)

Change-Id: I0dcc1623fafa7559edc8caf0b2b8db0817b721bc