Introduce buffer limit for dropbox entry
When logging the ECC diagnostic data, ensure that we truncate the data
if it will potentially exceed the 1MB kernel limit imposed on binder
transactions. Allowing the exception to be thrown and caught may cause
test failures due to the dropbox entry not being written (i.e. for the
logcat).

Bug: 441319428
Flag: EXEMPT bug/test fix
Test: atest TelephonyManagerTest#testPersistEmergencyCallDiagnosticData
Change-Id: I6299ca94a93018b3ddd3795e38b832d2091b3a6b