Include excludes.txt to fix presubmit crash
The presubmit check for TeleServiceTests is failing with an
`IllegalArgumentException` because the test runner,
`AndroidJUnitRunner`, is unable to find and read the `excludes.txt` file
at `/data/data/com.android.phone/excludes.txt`.

This crash is triggered when a change related to the Telephony provider
and the `UPDATE_APP_OPS_STATS` permission is submitted.

This commit adds a build rule to package an empty `excludes.txt` file
into the test APK, resolving the test runner crash and allowing the
presubmit tests to run as expected.

Bug: 443927435
Test: m aosp_arm64-trunk_staging-userdebug
Flag: EXEMPT BUGFIX
Change-Id: I18f725944bc7945f6e96c398dfa0965bc4ad8757